printer/requirements.txt
# Printer Service Dependencies
# Python 3.8.10 Compatible (32-bit and 64-bit)
#
# IMPORTANT: For old Windows compatibility, use 32-bit Python 3.8.10
# Download from: https://www.python.org/ftp/python/3.8.10/python-3.8.10.exe
# All packages below support both 32-bit and 64-bit architectures

# Windows Printing Support (tirada.py)
# Only install on Windows systems
# Note: pywin32 has separate wheels for 32-bit (win32) and 64-bit (amd64)
pywin32>=305; sys_platform == 'win32'

# ESC/POS Printer Support (recibo_*.py)
python-escpos>=2.2.0,<4.0.0
pyserial>=3.5,<4.0

# Image Processing
Pillow>=8.0.0,<11.0.0

# Network/API
requests>=2.25.0,<3.0.0

# Optional: For better serial port detection
pyusb>=1.2.1,<2.0.0
